requirements.txt
python-dotenv
requests
beautifulsoup4
typer[all]
rich